requirements.txt
Babel==2.0
cffi==1.2.1
cryptography==1.0.1
debtcollector==0.8.0
dnspython==1.12.0
enum34==1.0.4
eventlet==0.17.4
funcsigs==0.4
futures==3.0.3
greenlet==0.4.9
idna==2.0
ip-associations-python-novaclient-ext==0.1
ipaddress==1.0.14
iso8601==0.1.10
keyring==5.4
mock==1.3.0
monotonic==0.3
msgpack-python==0.4.6
ndg-httpsclient==0.4.0
netaddr==0.7.18
netifaces==0.10.4
os-diskconfig-python-novaclient-ext==0.1.2
os-networksv2-python-novaclient-ext==0.25
os-virtual-interfacesv2-python-novaclient-ext==0.19
oslo.config==2.4.0
oslo.i18n==2.6.0
oslo.serialization==1.9.0
oslo.utils==2.5.0
PasteDeploy==1.5.2
pbr==1.8.0
prettytable==0.7.2
psycopg2==2.5
pyasn1==0.1.8
pycparser==2.14
PyECLib==1.0.7
pyOpenSSL==0.15.1
pyrax==1.9.5
python-dateutil==2.4.2
python-keystoneclient==1.7.1
python-novaclient==2.30.0
python-swiftclient==2.6.0
pytz==2015.6
rackspace-auth-openstack==1.3
rackspace-novaclient==1.5
rax-default-network-flags-python-novaclient-ext==0.3.2
rax-scheduled-images-python-novaclient-ext==0.3.1
requests==2.7.0
simplejson==3.8.0
six==1.9.0
stevedore==1.8.0
git+https://github.com/openstack/swift/commit/51819f1179db246b3d2b9c6fd660e0f81a383db5#egg=swift
wheel==0.24.0
wrapt==1.10.5
xattr==0.7.8
